Controle de nivel de acesso de usuario realizado com sucesso!

+++ assets/view/pages/produto/listarProduto/listarProduto.php
@@ -1,7 +1,20 @@
 <?php
     require_once __DIR__ . "/../../../../controller/produtoController.php";
+    require_once __DIR__ . "/../../../../controller/userController.php";
+    require_once __DIR__ . "/../../../../model/utils.php"; 
+    session_start();
+
     $controler = new ControladorProdutos();
+    $user = $_SESSION['user'];
+    $tipo_usuario = $user->getTipoUsuario();
+
+    if(!isset($user)){
+        header("Location: index.php");
+    }
+
+    $podeGerenciarProdutos = verificarPermissao($tipo_usuario, 'gerenciar_produtos');
 ?>
+    
 <!DOCTYPE html>
 <html lang="pt-BR">
 <head>
@@ -19,21 +32,20 @@
         </section>
         <section class="user-info">
             <a href="../../principal.php" class="home-btn">Home</a>
-            <!-- <span><?php echo htmlspecialchars($user['nome']); ?></span> -->
             <a href="logout.php" class="logout-btn">Sair</a>
         </section>
     </section>
 </header>
 <main>
     <h1>Listar Produtos</h1>
-    <section class="search">
-        <form action="" method="POST">
+        <section class="search">
             <input type="text" name="search" placeholder="Pesquisar produtos...">
             <button type="submit">Pesquisar</button>
-        </form>
-    </section>
-    <section class="add-product">
-        <a href="../cadastroProduto/cadProduto.php" class="add-product-btn">Cadastrar Novo Produto</a>
+            <section class="add-product">
+            <?php if ($podeGerenciarProdutos): ?>
+                <a href="../cadastroProduto/cadProduto.php" class="add-product-btn">Cadastrar Novo Produto</a>
+                <?php endif; ?>
+            </section>
     </section>
     <table>
         <thead>
@@ -43,20 +55,22 @@
                 <th>Categoria</th>
                 <th>Unidade de Medida</th>
                 <th>Data de Criação</th>
+                <?php if ($podeGerenciarProdutos): ?>
                 <th colspan="2">Ações</th>
+                <?php endif; ?>
             </tr>
         </thead>
         <tbody>
             <?php
             $produtos = $controler->verProdutos();
             //TERMINAR DEPOIS
-            if ($_SERVER["REQUEST_METHOD"] == "POST"){
-                var_dump($produtos);
-                $like = $search = $_POST['search'];
+            //if ($_SERVER["REQUEST_METHOD"] == "POST"){
+              //  var_dump($produtos);
+                //$like = $search = $_POST['search'];
                 // if($search != ""){
                     
                 // }
-            }
+            //}
             if ($controler->verProdutos()) {
                 foreach ($produtos as $produto) {
                     echo "<tr>";
@@ -65,12 +79,14 @@
                     echo "<td>{$produto->getCategoria()}</td>";
                     echo "<td>{$produto->getUn()}</td>";
                     echo "<td>{$produto->getDtCriacao()}</td>";
-                    echo "<td><a href='../editarProduto/editProduto.php?id={$produto->getId()}'>Editar</a></td>";
-                    echo "<td><a href='../deleteProduto/delProduto.php?id={$produto->getId()}'>Deletar</a></td>";
+                    if ($podeGerenciarProdutos) {
+                        echo "<td> <a href='../editarProduto/editProduto.php?id={$produto->getId()}'>Editar</a> </td>";
+                        echo "<td> <a href='../deleteProduto/delProduto.php?id={$produto->getId()}'>Deletar</a> </td>";
+                    }
                     echo "</tr>";
                 }
             } else {
-                echo "<tr><td colspan='7'>Nenhum produto encontrado.</td></tr>";
+                echo "<tr></tr>";
             }
             ?>
         </tbody>
